Rename file, add waits
